+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Unreleased
 * Resolver plugins now allow more types of requires for hot reloading.
 * Raw routes are wrapped in main.
 
-### Removed (considerd non breaking)
+### Removed (considered non breaking)
 * Removed the webpackAliasOptions and monoRepo config sections from site.config. These were unused in any known repo and added complexity.
 
 6.0.0 - (July 30, 2019)
