+++ tests/config/webpack/resolve/DirectorySwitcherPlugin.test.js
@@ -1,6 +1,6 @@
 const path = require('path');
 
-const DirectorySwitcherPlugin = require('../../../plugin/resolve/DirectorySwitcherPlugin');
+const DirectorySwitcherPlugin = require('../../../../config/webpack/resolve/DirectorySwitcherPlugin');
 
 describe('DirectorySwitcherPlugin', () => {
   it('generates the dir structure', () => {
