+++ tests/config/webpack/resolve/LocalPackageAliasPlugin.test.js
@@ -1,6 +1,6 @@
 const path = require('path');
 
-const LocalPackageAliasPlugin = require('../../../config/webpack/resolve/LocalPackageAliasPlugin');
+const LocalPackageAliasPlugin = require('../../../../config/webpack/resolve/LocalPackageAliasPlugin');
 
 describe('LocalPackageAliasPlugin', () => {
   it('generates the alias structure', () => {
