+++ tests/config/webpack/resolve/packages/test/package.json
